Refactor pcapng: move ProcessInformation to its own file
