+++ src/pcapng.rs
@@ -53,6 +53,7 @@ mod enhanced_packet;
 mod interface_description;
 mod interface_statistics;
 mod name_resolution;
+mod process_information;
 mod section_header;
 mod simple_packet;
 mod systemd_journal_export;
@@ -62,6 +63,7 @@ pub use enhanced_packet::*;
 pub use interface_description::*;
 pub use interface_statistics::*;
 pub use name_resolution::*;
+pub use process_information::*;
 pub use section_header::*;
 pub use simple_packet::*;
 pub use systemd_journal_export::*;
@@ -355,45 +357,6 @@ impl<'a> CustomBlock<'a> {
     }
 }
 
-#[derive(Debug)]
-pub struct ProcessInformationBlock<'a> {
-    pub block_type: u32,
-    pub block_len1: u32,
-    pub process_id: u32,
-    pub options: Vec<PcapNGOption<'a>>,
-    pub block_len2: u32,
-}
-
-impl<'a, En: PcapEndianness> PcapNGBlockParser<'a, En, ProcessInformationBlock<'a>>
-    for ProcessInformationBlock<'a>
-{
-    const MAGIC: u32 = PIB_MAGIC;
-    const HDR_SZ: usize = 4;
-
-    fn inner_parse<E: ParseError<&'a [u8]>>(
-        block_type: u32,
-        block_len1: u32,
-        i: &'a [u8],
-        block_len2: u32,
-    ) -> IResult<&'a [u8], ProcessInformationBlock<'a>, E> {
-        // caller function already tested header type(magic) and length
-        // read options
-        let (i, process_id) = En::parse_u32(i)?;
-        let (i, options) = opt_parse_options::<En, E>(i, (block_len1 - 4) as usize, 12)?;
-        if block_len2 != block_len1 {
-            return Err(Err::Error(E::from_error_kind(i, ErrorKind::Verify)));
-        }
-        let block = ProcessInformationBlock {
-            block_type,
-            block_len1,
-            process_id,
-            options,
-            block_len2,
-        };
-        Ok((i, block))
-    }
-}
-
 /// Unknown block (magic not recognized, or not yet implemented)
 #[derive(Debug)]
 pub struct UnknownBlock<'a> {
@@ -630,20 +593,6 @@ pub fn parse_dcb_le(i: &[u8]) -> IResult<&[u8], CustomBlock, PcapError<&[u8]>> {
 #[inline]
 pub fn parse_dcb_be(i: &[u8]) -> IResult<&[u8], CustomBlock, PcapError<&[u8]>> {
     ng_block_parser::<DCBParser, PcapBE, _, _>()(i)
-}
-
-#[inline]
-pub fn parse_processinformationblock_le(
-    i: &[u8],
-) -> IResult<&[u8], ProcessInformationBlock, PcapError<&[u8]>> {
-    ng_block_parser::<ProcessInformationBlock, PcapLE, _, _>()(i)
-}
-
-#[inline]
-pub fn parse_processinformationblock_be(
-    i: &[u8],
-) -> IResult<&[u8], ProcessInformationBlock, PcapError<&[u8]>> {
-    ng_block_parser::<ProcessInformationBlock, PcapBE, _, _>()(i)
 }
 
 /// Parse an unknown block (little-endian)
